Add onclick to assBtn - use setNewBook (useState)
In your React component responsible for adding new books, add functionality for "Add new" button click:

Import Redux useDispatch.
Set the data inputs in the local React state (set title and author - NOTE: categories and comments are NOT part of this step).
Takes the inputs from the state and generate a unique id and add them to an object.
Once your new book object is ready to be submitted to Redux store, dispatch a corresponding action (import Action Creator that you created before and use it here).